Added TOTPURIHelper.cs for refactoring
